+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.1.0] - 2021-10-04
+
+- Fixed `accordion-item` block type markup bug
+- Added **Opened** toggle field in `accordion-item` block type for show the accordion item contents as default
+- Added body border setting fields in `accordion-item` block type
+- Improves in `fa-icon` block type
+- Added **Block level element type** toggle field in `fa-icon` for display it as an _Inline Element_ or _Block Element_
+- Fixed `border.js` helper bug
+- Removed `parent` attribute in `fa-icon` block type
+- Added new translations in `es_ES` language
+- Updated `readme.txt`
+
 ## [1.0.1] - 2021-09-29
 
 - Fix fa-icon block type